Refactor value printing transformations to support C and handle object ownership better.
Prior implementation was very dependent on C++ overloads.
Moving the implementation to a single C-function allows for C (and hopefully all languages that support a C calling convention to print values).
Additionally, previous behavior was to copy all non-pod objects into the Value, which is unnecessary in many cases.
This would often require two object’s to be constructed, and printing to fail if a copy constructor was unavailable.

Should also reduce startup time as there is much less to parse in RuntimePrintValue.h.