Fix AllocatedValue for cases when constructor or cling::printValue has thrown/failed.
Previous implementation was both non-conforming and could easily lead to a crash.
By delivering a memory as a char buffer, and allowing the client to write into the -1 slot,
ValuePrinterSynthesizer can mark that construction has succeeded and the destructor can be called.

Additionally this change hides most implementation details of AllocatedValue from cling::Value.